+++ Lib/test/test_unittest/__main__.py
@@ -1,0 +1,4 @@
+from . import load_tests
+import unittest
+
+unittest.main()
