+++ Lib/test/test_unittest/dummy.py
