+++ Lib/test/test_unittest/test_assertions.py
