+++ Lib/test/test_unittest/testmock/__init__.py
@@ -6,20 +6,12 @@ import unittest
 here = os.path.dirname(__file__)
 loader = unittest.defaultTestLoader
 
-def suite():
+def load_tests(*args):
     suite = unittest.TestSuite()
     for fn in os.listdir(here):
         if fn.startswith("test") and fn.endswith(".py"):
-            modname = "unittest.test." + fn[:-3]
-            try:
-                __import__(modname)
-            except unittest.SkipTest:
-                continue
+            modname = "test.test_unittest.testmock." + fn[:-3]
+            __import__(modname)
             module = sys.modules[modname]
             suite.addTest(loader.loadTestsFromModule(module))
-    suite.addTest(loader.loadTestsFromName('unittest.test.testmock'))
     return suite
-
-
-if __name__ == "__main__":
-    unittest.main(defaultTest="suite")
