+++ Lib/test/test_unittest/testmock/support.py
