+++ Lib/test/test_unittest/testmock/testhelpers.py
