+++ Lib/test/test_unittest/testmock/testmagicmethods.py
