+++ Lib/test/test_unittest/testmock/testsealable.py
