+++ PCbuild/lib.pyproj
@@ -1491,33 +1491,33 @@
     <Compile Include="unittest\runner.py" />
     <Compile Include="unittest\signals.py" />
     <Compile Include="unittest\suite.py" />
-    <Compile Include="unittest\test\dummy.py" />
-    <Compile Include="unittest\test\support.py" />
-    <Compile Include="unittest\test\testmock\support.py" />
-    <Compile Include="unittest\test\testmock\testcallable.py" />
-    <Compile Include="unittest\test\testmock\testhelpers.py" />
-    <Compile Include="unittest\test\testmock\testmagicmethods.py" />
-    <Compile Include="unittest\test\testmock\testmock.py" />
-    <Compile Include="unittest\test\testmock\testpatch.py" />
-    <Compile Include="unittest\test\testmock\testsentinel.py" />
-    <Compile Include="unittest\test\testmock\testwith.py" />
-    <Compile Include="unittest\test\testmock\__init__.py" />
-    <Compile Include="unittest\test\testmock\__main__.py" />
-    <Compile Include="unittest\test\test_assertions.py" />
-    <Compile Include="unittest\test\test_break.py" />
-    <Compile Include="unittest\test\test_case.py" />
-    <Compile Include="unittest\test\test_discovery.py" />
-    <Compile Include="unittest\test\test_functiontestcase.py" />
-    <Compile Include="unittest\test\test_loader.py" />
-    <Compile Include="unittest\test\test_program.py" />
-    <Compile Include="unittest\test\test_result.py" />
-    <Compile Include="unittest\test\test_runner.py" />
-    <Compile Include="unittest\test\test_setups.py" />
-    <Compile Include="unittest\test\test_skipping.py" />
-    <Compile Include="unittest\test\test_suite.py" />
-    <Compile Include="unittest\test\_test_warnings.py" />
-    <Compile Include="unittest\test\__init__.py" />
-    <Compile Include="unittest\test\__main__.py" />
+    <Compile Include="test\test_unittest\dummy.py" />
+    <Compile Include="test\test_unittest\support.py" />
+    <Compile Include="test\test_unittest\testmock\support.py" />
+    <Compile Include="test\test_unittest\testmock\testcallable.py" />
+    <Compile Include="test\test_unittest\testmock\testhelpers.py" />
+    <Compile Include="test\test_unittest\testmock\testmagicmethods.py" />
+    <Compile Include="test\test_unittest\testmock\testmock.py" />
+    <Compile Include="test\test_unittest\testmock\testpatch.py" />
+    <Compile Include="test\test_unittest\testmock\testsentinel.py" />
+    <Compile Include="test\test_unittest\testmock\testwith.py" />
+    <Compile Include="test\test_unittest\testmock\__init__.py" />
+    <Compile Include="test\test_unittest\testmock\__main__.py" />
+    <Compile Include="test\test_unittest\test_assertions.py" />
+    <Compile Include="test\test_unittest\test_break.py" />
+    <Compile Include="test\test_unittest\test_case.py" />
+    <Compile Include="test\test_unittest\test_discovery.py" />
+    <Compile Include="test\test_unittest\test_functiontestcase.py" />
+    <Compile Include="test\test_unittest\test_loader.py" />
+    <Compile Include="test\test_unittest\test_program.py" />
+    <Compile Include="test\test_unittest\test_result.py" />
+    <Compile Include="test\test_unittest\test_runner.py" />
+    <Compile Include="test\test_unittest\test_setups.py" />
+    <Compile Include="test\test_unittest\test_skipping.py" />
+    <Compile Include="test\test_unittest\test_suite.py" />
+    <Compile Include="test\test_unittest\_test_warnings.py" />
+    <Compile Include="test\test_unittest\__init__.py" />
+    <Compile Include="test\test_unittest\__main__.py" />
     <Compile Include="unittest\util.py" />
     <Compile Include="unittest\__init__.py" />
     <Compile Include="unittest\__main__.py" />
@@ -1804,6 +1804,8 @@
     <Folder Include="test\test_json" />
     <Folder Include="test\test_peg_generator" />
     <Folder Include="test\test_tools" />
+    <Folder Include="test\test_unittest" />
+    <Folder Include="test\test_unittest\testmock" />
     <Folder Include="test\test_warnings" />
     <Folder Include="test\test_warnings\data" />
     <Folder Include="test\tracedmodules" />
@@ -1813,8 +1815,6 @@
     <Folder Include="tkinter\test\test_ttk" />
     <Folder Include="turtledemo" />
     <Folder Include="unittest" />
-    <Folder Include="unittest\test" />
-    <Folder Include="unittest\test\testmock" />
     <Folder Include="urllib" />
     <Folder Include="venv" />
     <Folder Include="wsgiref" />
