+++ Tools/wasm/wasm_assets.py
@@ -112,7 +112,6 @@ OMIT_MODULE_FILES = {
 # regression test sub directories
 OMIT_SUBDIRS = (
     "tkinter/test/",
-    "unittest/test/",
 )
 
 def get_builddir(args: argparse.Namespace) -> pathlib.Path:
